saqlain2022skipq/Training_Sprint_7/Design_Problem_7/requirements.txt
aws-cdk-lib==2.76.0
constructs>=10.0.0,<11.0.0
requests==2.28.2
